Rename subprojects of micronaut-camunda-bpm to ...-feature and ...-example
