+++ micronaut-camunda-bpm-example/Dockerfile
