+++ micronaut-camunda-bpm-example/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name="micronaut-camunda-bpm-example"
