+++ micronaut-camunda-bpm-example/src/main/java/info/novatec/micronaut/camunda/bpm/example/Application.java
@@ -1,0 +1,10 @@
+package info.novatec.micronaut.camunda.bpm.example;
+
+import io.micronaut.runtime.Micronaut;
+
+public class Application {
+
+    public static void main(String[] args) {
+        Micronaut.run(Application.class);
+    }
+}
