+++ micronaut-camunda-bpm-example/src/main/java/info/novatec/micronaut/camunda/bpm/example/GreetingService.java
@@ -1,0 +1,10 @@
+package info.novatec.micronaut.camunda.bpm.example;
+
+import javax.inject.Singleton;
+
+@Singleton
+public class GreetingService {
+    public String getGreeting(String name) {
+        return "Hello " + name;
+    }
+}
