+++ micronaut-camunda-bpm-example/src/main/resources/application.yml
