+++ micronaut-camunda-bpm-example/src/main/resources/bpm/helloworld.bpmn
