+++ micronaut-camunda-bpm-example/src/test/java/info/novatec/micronaut/camunda/bpm/example/LoggerFactory.java
@@ -1,0 +1,18 @@
+package info.novatec.micronaut.camunda.bpm.example;
+
+import io.micronaut.context.annotation.Factory;
+import io.micronaut.context.annotation.Replaces;
+import org.mockito.Mockito;
+
+import javax.inject.Named;
+import javax.inject.Singleton;
+
+@Factory
+public class LoggerFactory {
+    @Singleton
+    @Replaces(LoggerDelegate.class)
+    @Named("loggerDelegate")
+    public LoggerDelegate loggerDelegate() {
+        return Mockito.mock(LoggerDelegate.class);
+    }
+}
