+++ micronaut-camunda-bpm-feature/micronaut-cli.yml
@@ -1,0 +1,5 @@
+profile: service
+defaultPackage: info.novatec.micronaut.camunda.bpm.feature
+---
+testFramework: junit
+sourceLanguage: java
