+++ micronaut-camunda-bpm-feature/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name="micronaut-camunda-bpm-feature"
