+++ micronaut-camunda-bpm-feature/src/test/resources/empty.bpmn
