--- micronaut-camunda-example/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name="micronaut-camunda-example"
