--- micronaut-camunda-example/src/main/java/info/novatec/micronaut/camunda/example/Application.java
@@ -1,10 +1,0 @@
-package info.novatec.micronaut.camunda.example;
-
-import io.micronaut.runtime.Micronaut;
-
-public class Application {
-
-    public static void main(String[] args) {
-        Micronaut.run(Application.class);
-    }
-}
