--- micronaut-camunda-example/src/main/java/info/novatec/micronaut/camunda/example/GreetingService.java
@@ -1,10 +1,0 @@
-package info.novatec.micronaut.camunda.example;
-
-import javax.inject.Singleton;
-
-@Singleton
-public class GreetingService {
-    public String getGreeting(String name) {
-        return "Hello " + name;
-    }
-}
