--- micronaut-camunda-feature/build.gradle
@@ -1,56 +1,0 @@
-plugins {
-    id "net.ltgt.apt-eclipse" version "0.21"
-    id "application"
-}
-
-version "0.1-SNAPSHOT"
-group "info.novatec.micronaut.camunda.feature"
-
-configurations {
-    // for dependencies that are needed for development only
-    developmentOnly 
-}
-
-dependencies {
-    annotationProcessor platform("io.micronaut:micronaut-bom:$micronautVersion")
-    annotationProcessor "io.micronaut:micronaut-inject-java"
-    annotationProcessor "io.micronaut:micronaut-validation"
-    implementation platform("io.micronaut:micronaut-bom:$micronautVersion")
-    implementation "io.micronaut:micronaut-inject"
-    implementation "io.micronaut:micronaut-validation"
-    implementation "io.micronaut:micronaut-runtime"
-    implementation "io.micronaut:micronaut-http-server-netty"
-    implementation "io.micronaut:micronaut-http-client"
-    implementation "io.micronaut:micronaut-cli"
-    implementation "org.camunda.bpm:camunda-engine:7.12.0"
-    testImplementation "com.h2database:h2:1.4.200"
-    runtimeOnly "ch.qos.logback:logback-classic:1.2.3"
-    testAnnotationProcessor platform("io.micronaut:micronaut-bom:$micronautVersion")
-    testAnnotationProcessor "io.micronaut:micronaut-inject-java"
-    testImplementation platform("io.micronaut:micronaut-bom:$micronautVersion")
-    testImplementation "org.junit.jupiter:junit-jupiter-api"
-    testImplementation "io.micronaut.test:micronaut-test-junit5"
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine"
-}
-
-test.classpath += configurations.developmentOnly
-
-mainClassName = "info.novatec.micronaut.camunda.example.Application"
-// use JUnit 5 platform
-test {
-    useJUnitPlatform()
-}
-tasks.withType(JavaCompile){
-    options.encoding = "UTF-8"
-    options.compilerArgs.add('-parameters')
-}
-
-tasks.withType(JavaExec) {
-    classpath += configurations.developmentOnly
-    jvmArgs('-noverify', '-XX:TieredStopAtLevel=1', '-Dcom.sun.management.jmxremote')
-}
-
-java {
-    withJavadocJar()
-    withSourcesJar()
-}
