--- micronaut-camunda-feature/micronaut-cli.yml
@@ -1,5 +1,0 @@
-profile: service
-defaultPackage: info.novatec.micronaut.camunda.feature
----
-testFramework: junit
-sourceLanguage: java
